+++ src/js/preprocess.js
@@ -73,8 +73,10 @@ if (supportsOffscreenCanvas) {
         const channel = new MessageChannel();
         channel.port1.onmessage = ({ data }) => {
           channel.port1.close();
-          preProcessWorker.terminate();
-          preProcessWorker = null;
+          if (preProcessWorker) {
+            preProcessWorker.terminate();
+            preProcessWorker = null;
+          }
           canvasMain.width = width;
           canvasMain.height = height;
           ctxCanvasMain.putImageData(data.result, 0, 0);
