+++ src/js/svgo.js
@@ -32,8 +32,10 @@ const optimizeSVG = async (svg) => {
     const channel = new MessageChannel();
     channel.port1.onmessage = ({ data }) => {
       channel.port1.close();
-      optimizeSVGWorker.terminate();
-      optimizeSVGWorker = null;
+      if (optimizeSVGWorker) {
+        optimizeSVGWorker.terminate();
+        optimizeSVGWorker = null;
+      }
       resolve(data.result);
     };
